requirements.txt
numpy
pandas
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
PyYaml>=5.4
dvc>=2.5